# Conflicts:
#	RevenueSharingInvest.API/Controllers/BusinessController.cs

+++ RevenueSharingInvest.API/Controllers/BusinessController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using RevenueSharingInvest.Business.Exceptions;
 using RevenueSharingInvest.Business.Models.Constant;
 using RevenueSharingInvest.Business.Services;
 using RevenueSharingInvest.Data.Models.Constants;
@@ -35,9 +36,14 @@ namespace RevenueSharingInvest.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateBusiness([FromForm] CreateUpdateBusinessDTO businessDTO, [FromQuery] List<string> fieldIdList)
+        public async Task<IActionResult> CreateBusiness([FromForm] CreateBusinessDTO businessDTO, [FromQuery] List<string> fieldIdList)
         {
             ThisUserObj userInfo = await GetThisUserInfo(HttpContext);
+
+            if (!userInfo.businessId.Equals(""))
+            {
+                throw new CreateObjectException("This BUSINESS_MANAGER has a business already!!!");
+            }
 
             if(userInfo.roleId.Equals(userInfo.businessManagerRoleId))
             {
@@ -54,14 +60,14 @@ namespace RevenueSharingInvest.API.Controllers
         {
             ThisUserObj userInfo = await GetThisUserInfo(HttpContext);
 
-            if((userInfo.roleId.Equals(userInfo.investorRoleId) && userInfo.investorId != null) || userInfo.roleId.Equals(userInfo.adminRoleId))
+            if((userInfo.roleId.Equals(userInfo.investorRoleId) && !userInfo.investorId.Equals("")) || userInfo.roleId.Equals(userInfo.adminRoleId))
             {
                 var result = new AllBusinessDTO();
                 result = await _businessService.GetAllBusiness(pageIndex, pageSize, orderBy, order, userInfo.roleId);
                 return Ok(result);
             }
 
-            return StatusCode((int)HttpStatusCode.Forbidden, "You Do Not Have Permission To Perform This Action!!");
+            return StatusCode((int)HttpStatusCode.Forbidden, "Only user with role ADMIN and INVESTOR can perform this action!!");
         }
 
         [HttpGet]
@@ -71,7 +77,7 @@ namespace RevenueSharingInvest.API.Controllers
             ThisUserObj userInfo = await GetThisUserInfo(HttpContext);
             GetBusinessDTO dto = new GetBusinessDTO();
 
-            if((userInfo.roleId.Equals(userInfo.businessManagerRoleId) || userInfo.roleId.Equals(userInfo.projectManagerRoleId)) && userInfo.businessId != null)
+            if((userInfo.roleId.Equals(userInfo.businessManagerRoleId) || userInfo.roleId.Equals(userInfo.projectManagerRoleId)) && !userInfo.businessId.Equals(""))
             {
                 dto = await _businessService.GetBusinessById(Guid.Parse(userInfo.businessId));
                 return Ok(dto);
@@ -82,11 +88,16 @@ namespace RevenueSharingInvest.API.Controllers
 
         [HttpPut]
         [Route("{id}")]
-        public async Task<IActionResult> UpdateBusiness([FromForm] CreateUpdateBusinessDTO businessDTO)
+        public async Task<IActionResult> UpdateBusiness([FromForm] UpdateBusinessDTO businessDTO)
         {
             ThisUserObj userInfo = await GetThisUserInfo(HttpContext);
+
             if (userInfo.roleId.Equals(userInfo.businessManagerRoleId))
             {
+                if (userInfo.businessId.Equals(""))
+                {
+                    throw new System.UnauthorizedAccessException("You Have To Create Business First!!");
+                }
                 var result = await _businessService.UpdateBusiness(businessDTO, Guid.Parse(userInfo.businessId));
                 return Ok(result);
             }
@@ -100,15 +111,19 @@ namespace RevenueSharingInvest.API.Controllers
         {
             ThisUserObj userInfo = await GetThisUserInfo(HttpContext);
 
+            GetBusinessDTO businessDTO = await _businessService.GetBusinessById(id);
+
+            if(businessDTO == null)
+            {
+                throw new Business.Exceptions.NotFoundException("No Business With This ID Found!!");
+            }
+
             if (userInfo.roleId.Equals(userInfo.adminRoleId))
             {
                 var result = await _businessService.DeleteBusinessById(id);
                 return Ok(result);
             }
-
             return StatusCode((int)HttpStatusCode.Forbidden, "You Do Not Have Permission To Access This Business!!");
-
-
         }
 
         [HttpDelete]
@@ -141,6 +156,10 @@ namespace RevenueSharingInvest.API.Controllers
             {
                 currentUser.businessId = userDTO.business.id;
             }
+            else
+            {
+                currentUser.businessId = "";
+            }
 
             foreach(RoleDTO role in roleList)
             {
@@ -148,15 +167,15 @@ namespace RevenueSharingInvest.API.Controllers
                 {
                     currentUser.adminRoleId = role.id;
                 }                
-                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(1)))
+                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(3)))
                 {
                     currentUser.investorRoleId = role.id;
                 }                
-                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(2)))
+                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(1)))
                 {
                     currentUser.businessManagerRoleId = role.id;
                 }                
-                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(3)))
+                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(2)))
                 {
                     currentUser.projectManagerRoleId = role.id;
                 }
