+++ RevenueSharingInvest.API/Controllers/ProjectController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using RevenueSharingInvest.Business.Exceptions;
 using RevenueSharingInvest.Business.Models.Constant;
 using RevenueSharingInvest.Business.Services;
 using RevenueSharingInvest.Data.Models.DTOs;
@@ -57,8 +58,6 @@ namespace RevenueSharingInvest.API.Controllers
                 var result = await _projectService.CreateProject(projectDTO);
                 return Ok(result);
             }
-
-
             return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
         }
 
@@ -77,44 +76,37 @@ namespace RevenueSharingInvest.API.Controllers
             string status
             )
         {
-
             ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
 
             if (countOnly)
             {
                 var countResult = new ProjectCountDTO();
-
-                if(currentUser.roleId != null)
+                if(!currentUser.roleId.Equals(""))
                 {
                     RoleDTO roleDTO = await _roleService.GetRoleById(Guid.Parse(currentUser.roleId));
-                    countResult = await _projectService.CountProjects(businessId, managerId, areaId, fieldId, investorId, name, status.ToString(), roleDTO.name);
+                    countResult = await _projectService.CountProjects(businessId, managerId, areaId, fieldId, investorId, name, status, roleDTO.name);
                     return Ok(countResult);
                 } else
                 {
-                    countResult = await _projectService.CountProjects(businessId, managerId, areaId, fieldId, investorId, name, status.ToString(), "GUEST");
+                    countResult = await _projectService.CountProjects(businessId, managerId, areaId, fieldId, investorId, name, status, "GUEST");
                     return Ok(countResult);
                 }
-
             }
             else
             {
-
                 var resultProjectList = new AllProjectDTO();
-
-                if(currentUser.roleId != null)
+                if(!currentUser.roleId.Equals(""))
                 {
                     RoleDTO roleDTO = await _roleService.GetRoleById(Guid.Parse(currentUser.roleId));
 
-                    resultProjectList = await _projectService.GetAllProjects(pageIndex, pageSize, businessId, managerId, areaId, fieldId, investorId, name, status.ToString(), roleDTO.name);
+                    resultProjectList = await _projectService.GetAllProjects(pageIndex, pageSize, businessId, managerId, areaId, fieldId, investorId, name, status, roleDTO.name);
                     return Ok(resultProjectList);
                 } else
                 {
-                    resultProjectList = await _projectService.GetAllProjects(pageIndex, pageSize, businessId, managerId, areaId, fieldId, investorId, name, status.ToString(), "GUEST");
+                    resultProjectList = await _projectService.GetAllProjects(pageIndex, pageSize, businessId, managerId, areaId, fieldId, investorId, name, status, "GUEST");
                     return Ok(resultProjectList);
                 }
-
             }           
-
             return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
         }
 
@@ -134,9 +126,9 @@ namespace RevenueSharingInvest.API.Controllers
 
             } else if (currentUser.roleId.Equals(currentUser.businessManagerRoleId))
             {
-                if (currentUser.businessId == null || currentUser.businessId == "")
+                if (currentUser.businessId.Equals(""))
                 {
-                    throw new UnauthorizedAccessException("You Don't Have Permission Perform This Action!!");
+                    throw new System.UnauthorizedAccessException("You Don't Have Permission Perform This Action!!");
                 }
 
                 List<BusinessProjectDTO> projectList = await _projectService.GetBusinessProjectsToAuthor(Guid.Parse(currentUser.businessId));
@@ -150,9 +142,9 @@ namespace RevenueSharingInvest.API.Controllers
                 }
             } else if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
             {
-                if (currentUser.businessId == null || currentUser.businessId == "")
+                if (currentUser.businessId.Equals(""))
                 {
-                    throw new UnauthorizedAccessException("You Don't Have Permission Perform This Action!!");
+                    throw new System.UnauthorizedAccessException("You Don't Have Permission Perform This Action!!");
                 }
 
                 List<BusinessProjectDTO> projectList = await _projectService.GetBusinessProjectsToAuthor(Guid.Parse(currentUser.businessId));
@@ -185,7 +177,7 @@ namespace RevenueSharingInvest.API.Controllers
                     return Ok(dto);
                 } else
                 {
-                    throw new UnauthorizedAccessException("You Don't Have Permission Perform This Action!!");
+                    throw new System.UnauthorizedAccessException("You Don't Have Permission Perform This Action!!");
                 }
                 
             }
@@ -199,14 +191,34 @@ namespace RevenueSharingInvest.API.Controllers
         [Authorize]
         public async Task<IActionResult> UpdateProject([FromForm] CreateUpdateProjectDTO projectDTO, Guid id)
         {
-            //string userId = _httpContextAccessor.HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.SerialNumber).Value;
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
 
-            //if (await _authenticateService.CheckRoleForAction(userId, RoleEnum.BUSINESS_MANAGER.ToString()))
-            //{
-                var result = await _projectService.UpdateProject(projectDTO, id);
-                return Ok(result);
-            //}
-            //return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
+            if (currentUser.roleId.Equals(currentUser.businessManagerRoleId))
+            {
+                Data.Models.Entities.Business businessDTO = await _businessService.GetBusinessByProjectId(id);
+
+                GetProjectDTO project = await _projectService.GetProjectById(id);
+
+                if (businessDTO == null || project == null)
+                {
+                    throw new NotFoundException("no Project With This ID Found!!");
+                } else if(businessDTO.Id.Equals(currentUser.businessId))
+                {
+                    if(project.status.Equals(ProjectStatusEnum.DRAFT.ToString()) || project.status.Equals(ProjectStatusEnum.WAITING_FOR_APPROVED.ToString()))
+                    {
+                        var result = await _projectService.UpdateProject(projectDTO, id);
+                        return Ok(result);
+                    }   
+                }
+
+                
+            } else if (currentUser.roleId.Equals(currentUser.projectManagerRoleId))
+            {
+
+            }
+
+
+            return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
         }
 
         [HttpDelete]
@@ -214,14 +226,31 @@ namespace RevenueSharingInvest.API.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteProject(Guid id)
         {
-            //string userId = _httpContextAccessor.HttpContext.User.Claims.First(c => c.Type == ClaimTypes.SerialNumber).Value;
+            ThisUserObj currentUser = await GetThisUserInfo(HttpContext);
 
-            //if (await _authenticateService.CheckRoleForAction(userId, RoleEnum.ADMIN.ToString()) && await _authenticateService.CheckIdForAction(userId, id))
-            //{
-                var result = await _projectService.DeleteProjectById(id);
-                return Ok(result);
-            //}
-            //return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
+            if (currentUser.roleId.Equals(currentUser.adminRoleId))
+            {
+                GetProjectDTO project = await _projectService.GetProjectById(id);
+
+                if (!project.status.Equals(ProjectStatusEnum.DRAFT.ToString()))
+                {
+                    var result = await _projectService.DeleteProjectById(id);
+                    return Ok(result);
+                }
+            } else if (currentUser.roleId.Equals(currentUser.businessManagerRoleId))
+            {
+                GetProjectDTO project = await _projectService.GetProjectById(id);
+
+                if (project.status.Equals(ProjectStatusEnum.DRAFT.ToString()) 
+                    || project.status.Equals(ProjectStatusEnum.DENIED.ToString()) 
+                    || project.status.Equals(ProjectStatusEnum.CLOSED.ToString()))
+                {
+                    var result = await _projectService.DeleteProjectById(id);
+                    return Ok(result);
+                }
+            }
+
+            return StatusCode((int)HttpStatusCode.Forbidden, "You Don't Have Permission Perform This Action!!");
 
         }
 
@@ -252,7 +281,14 @@ namespace RevenueSharingInvest.API.Controllers
             GetUserDTO userDTO = await _userService.GetUserByEmail(currentUser.email);
 
             currentUser.roleId = userDTO.role.id;
-            currentUser.businessId = userDTO.business.id;
+            if (userDTO.business != null)
+            {
+                currentUser.businessId = userDTO.business.id;
+            }
+            else
+            {
+                currentUser.businessId = "";
+            }
 
             foreach (RoleDTO role in roleList)
             {
@@ -260,15 +296,15 @@ namespace RevenueSharingInvest.API.Controllers
                 {
                     currentUser.adminRoleId = role.id;
                 }
-                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(1)))
+                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(3)))
                 {
                     currentUser.investorRoleId = role.id;
                 }
-                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(2)))
+                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(1)))
                 {
                     currentUser.businessManagerRoleId = role.id;
                 }
-                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(3)))
+                if (role.name.Equals(Enum.GetNames(typeof(RoleEnum)).ElementAt(2)))
                 {
                     currentUser.projectManagerRoleId = role.id;
                 }
